Refactored and modularized structure
- Split app.py into core/routes, core/services, and core/utils
- Added main_routes.py, file_routes.py, image_routes.py for routing
- Improved image_upload.html UI and added Home button
- Prevented duplicate PDF conversions in file_details.html
- Updated requirements.txt and Docker setup
- Enhanced logging for debugging
