+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     ports:
       - "7777:7777"
     environment:
-    - OPENAI_API_KEY=${OPENAI_API_KEY}
-    - APP_SECRET_KEY=${APP_SECRET_KEY}
+      - OPENAI_API_KEY=${OPENAI_API_KEY}
+      - APP_SECRET_KEY=${APP_SECRET_KEY}
     volumes:
       - .:/usr/src/app
