+++ templates/index.html
@@ -31,29 +31,25 @@
             border-radius: 40px;
             box-shadow: rgba(0, 0, 0, 0.24) 0px 10px 20px,
             rgba(0, 0, 0, 0.28) 0px 6px 6px;
-
+        }
         form {
             display: flex;
             align-items: center;
             justify-content: center;
             flex-direction: column;
         }
-
         h1 {
             margin-bottom: 0;
-        }
         }
         .upload-files-list {
             list-style: none;
             padding: 0;
-
-        li {
+        }
+        .upload-files-list li {
             margin-bottom: 2px;
         }
-
-        a {
+        .upload-files-list a {
             color: white;
-        }
         }
         .drag-file-area {
             border: 2px dashed #76B900;
@@ -62,32 +58,27 @@
             padding: 30px 50px;
             width: 400px;
             text-align: center;
-            transition: border-style 0.5;
-
+            transition: border-style 0.5s;
+        }
         .upload-icon {
             font-size: 50px;
         }
-
-        h3 {
+        .drag-file-area h3 {
             font-size: 26px;
             margin: 15px 0;
         }
-
-        label {
+        .drag-file-area label {
             font-size: 19px;
-
+        }
         .delete-link {
             margin-left: 10px;
             cursor: pointer;
             vertical-align: middle;
         }
-
         .browse-files-text {
             color: #76B900;
             font-weight: bolder;
             cursor: pointer;
-        }
-        }
         }
         .browse-files span {
             position: relative;
@@ -181,7 +172,6 @@
             padding: 7.5px 50px;
             cursor: pointer;
         }
-
         .upload-icon {
             transition: transform 0.5s;
         }
@@ -191,14 +181,12 @@
         .solid-border {
             border-style: solid;
         }
-
         .area {
             background: #f7fff7;
             background: -webkit-linear-gradient(to left, #76B900, #f7fff7);
             width: 100vw;
             min-height: 100vh;
         }
-
         .circles {
             padding: 0;
             margin: 0;
@@ -211,7 +199,6 @@
             height: 100%;
             overflow: hidden;
         }
-
         .circles li {
             position: absolute;
             display: block;
@@ -222,14 +209,12 @@
             animation: animate 25s linear infinite;
             bottom: -150px;
         }
-
         .circles li:nth-child(1) {
             left: 25%;
             width: 80px;
             height: 80px;
             animation-delay: 0s;
         }
-
         .circles li:nth-child(2) {
             left: 10%;
             width: 20px;
@@ -237,14 +222,12 @@
             animation-delay: 2s;
             animation-duration: 12s;
         }
-
         .circles li:nth-child(3) {
             left: 70%;
             width: 20px;
             height: 20px;
             animation-delay: 4s;
         }
-
         .circles li:nth-child(4) {
             left: 40%;
             width: 60px;
@@ -252,28 +235,24 @@
             animation-delay: 0s;
             animation-duration: 18s;
         }
-
         .circles li:nth-child(5) {
             left: 65%;
             width: 20px;
             height: 20px;
             animation-delay: 0s;
         }
-
         .circles li:nth-child(6) {
             left: 75%;
             width: 110px;
             height: 110px;
             animation-delay: 3s;
         }
-
         .circles li:nth-child(7) {
             left: 35%;
             width: 150px;
             height: 150px;
             animation-delay: 7s;
         }
-
         .circles li:nth-child(8) {
             left: 50%;
             width: 25px;
@@ -281,7 +260,6 @@
             animation-delay: 15s;
             animation-duration: 45s;
         }
-
         .circles li:nth-child(9) {
             left: 20%;
             width: 15px;
@@ -289,7 +267,6 @@
             animation-delay: 2s;
             animation-duration: 35s;
         }
-
         .circles li:nth-child(10) {
             left: 85%;
             width: 150px;
@@ -297,14 +274,12 @@
             animation-delay: 0s;
             animation-duration: 11s;
         }
-
         @keyframes animate {
             0% {
                 transform: translateY(0) rotate(0deg);
                 opacity: 1;
                 border-radius: 0;
             }
-
             100% {
                 transform: translateY(-1000px) rotate(720deg);
                 opacity: 0;
@@ -331,7 +306,7 @@
 <div class="form-container">
     <div class="upload-files-container">
         <form
-                action="{{ url_for('upload_files') }}"
+                action="{{ url_for('main.upload_files') }}"
                 method="post"
                 enctype="multipart/form-data"
         >
@@ -339,9 +314,9 @@
             <ul class="upload-files-list">
                 {% for file in files %}
                     <li>
-                        <a href="{{ url_for('file_details', filename=file) }}">{{ file }}</a>
+                        <a href="{{ url_for('file.file_details', filename=file) }}">{{ file }}</a>
                         <!-- Delete button -->
-                        <a href="{{ url_for('delete_file', filename=file) }}" class="delete-link" title="Delete">
+                        <a href="{{ url_for('main.delete_file', filename=file) }}" class="delete-link" title="Delete">
                             <span class="material-icons" style="color: red;">delete</span>
                         </a>
                     </li>
@@ -375,7 +350,7 @@
             <button type="submit" class="upload-button">Upload</button>
             <p>or</p>
         </form>
-        <form action="{{ url_for('image_upload') }}" method="get" style="margin: 0;">
+        <form action="{{ url_for('image.image_upload') }}" method="get" style="margin: 0;">
             <button type="submit" class="upload-button" style="text-decoration: none;">Upload singular Images</button>
         </form>
     </div>
